created student Database Management system project
This is a simple database management system developed in c++.It utilizes file handling and shows effective implementation of class and object of programming language .This project will teach you how to add, list ,modify and delete records in file in c++ language. It is very simple to understand and the source code too is not that long ,just under 200 lines.